:art: 抽卡记录链接功能放到抽卡记录插件里
